pyproject.toml
[project]
name = "ai-news-aggregator"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12",
    "google-generativeai>=0.8.6",
    "openai>=2.30.0",
    "pydantic>=2.12.5",
    "requests>=2.31",
]
